# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication, Authorization, and Accounting Service with PostgreSQL-based RBAC",
        "title": "AAA Service API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v2/modules": {
            "get": {
                "description": "Get a list of all registered modules",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "List all registered modules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ModuleListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v2/modules/register": {
            "post": {
                "description": "Register a complete module with actions, roles, resources, and permissions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Register a new module",
                "parameters": [
                    {
                        "description": "Module registration request",
                        "name": "module",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.ModuleRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.ModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v2/modules/{service_name}": {
            "get": {
                "description": "Get detailed information about a registered module",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Get module information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "service_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.ModuleActionDefinition": {
            "description": "Action definition structure for module registration",
            "type": "object",
            "required": [
                "name",
                "resource"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Create a new user"
                },
                "name": {
                    "type": "string",
                    "example": "create_user"
                },
                "resource": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "requests.ModuleRegistrationRequest": {
            "description": "Request structure for registering a new module",
            "type": "object",
            "required": [
                "service_name",
                "version"
            ],
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.ModuleActionDefinition"
                    }
                },
                "resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.ModuleResourceDefinition"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.ModuleRoleDefinition"
                    }
                },
                "service_name": {
                    "type": "string",
                    "example": "user-service"
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "requests.ModuleResourceDefinition": {
            "description": "Resource definition structure for module registration",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"create\"",
                        " \"read\"",
                        " \"update\"",
                        " \"delete\"]"
                    ]
                },
                "description": {
                    "type": "string",
                    "example": "User resource"
                },
                "name": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "requests.ModuleRoleDefinition": {
            "description": "Role definition structure for module registration",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "User administrator role"
                },
                "name": {
                    "type": "string",
                    "example": "user_admin"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"user:create\"",
                        " \"user:read\"",
                        " \"user:update\"",
                        " \"user:delete\"]"
                    ]
                }
            }
        },
...
        "responses.ModuleDetailData": {
            "description": "Module detail data structure",
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "create",
                        "read",
                        "update",
                        "delete"
                    ]
                },
                "module": {
                    "$ref": "#/definitions/responses.ModuleInfo"
                },
                "resources": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user",
                        "role"
                    ]
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "admin",
                        "user"
                    ]
                },
                "service_id": {
                    "type": "string",
                    "example": "svc_123456"
                }
            }
        },
        "responses.ModuleDetailResponse": {
            "description": "Module detail response structure",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.ModuleDetailData"
                },
                "message": {
                    "type": "string",
                    "example": "Module details retrieved successfully"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456789"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "responses.ModuleInfo": {
            "description": "Module information structure",
            "type": "object",
            "properties": {
                "health": {
                    "type": "string",
                    "example": "healthy"
                },
                "last_seen": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "registered_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "service_name": {
                    "type": "string",
                    "example": "user-service"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "responses.ModuleListData": {
            "description": "Module list data structure",
            "type": "object",
            "properties": {
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ModuleInfo"
                    }
                }
            }
        },
        "responses.ModuleListResponse": {
            "description": "Module list response structure",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.ModuleListData"
                },
                "message": {
                    "type": "string",
                    "example": "Modules retrieved successfully"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456789"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}